Split devices tests out into a separate directory.
They were pretty big and diverse, just like devices.py/devices directory.
