+++ tests/devices_test/device_properties_test.py
@@ -1,7 +1,6 @@
 #!/usr/bin/python
 # vim:set fileencoding=utf-8
 
-import os
 import unittest
 from gi.repository import BlockDev as blockdev
 
@@ -16,24 +15,14 @@ from blivet.devices import BTRFSSnapShotDevice
 from blivet.devices import BTRFSSubVolumeDevice
 from blivet.devices import BTRFSVolumeDevice
 from blivet.devices import DiskDevice
-from blivet.devices import DiskFile
-from blivet.devices import LVMLogicalVolumeDevice
-from blivet.devices import LVMSnapShotDevice
-from blivet.devices import LVMThinPoolDevice
-from blivet.devices import LVMThinLogicalVolumeDevice
-from blivet.devices import LVMThinSnapShotDevice
-from blivet.devices import LVMVolumeGroupDevice
 from blivet.devices import MDBiosRaidArrayDevice
 from blivet.devices import MDContainerDevice
 from blivet.devices import MDRaidArrayDevice
-from blivet.devices import NetworkStorageDevice
 from blivet.devices import OpticalDevice
-from blivet.devices import PartitionDevice
 from blivet.devices import StorageDevice
 from blivet.devices import ParentList
 from blivet.devicelibs import btrfs
 from blivet.size import Size
-from blivet.util import sparsetmpfile
 
 from blivet.formats import getFormat
 
@@ -723,329 +712,5 @@ class BTRFSDeviceTestCase(DeviceStateTestCase):
         self.assertEqual(snap.dependsOn(vol), True)
         self.assertEqual(vol.dependsOn(snap), False)
 
-class LVMDeviceTest(unittest.TestCase):
-    def testLVMSnapShotDeviceInit(self):
-        pv = StorageDevice("pv1", fmt=blivet.formats.getFormat("lvmpv"),
-                           size=Size("1 GiB"))
-        vg = LVMVolumeGroupDevice("testvg", parents=[pv])
-        lv = LVMLogicalVolumeDevice("testlv", parents=[vg],
-                                    fmt=blivet.formats.getFormat("xfs"))
-
-        with self.assertRaisesRegexp(ValueError, "lvm snapshot devices require an origin lv"):
-            LVMSnapShotDevice("snap1", parents=[vg])
-
-        with self.assertRaisesRegexp(ValueError, "lvm snapshot origin volume must already exist"):
-            LVMSnapShotDevice("snap1", parents=[vg], origin=lv)
-
-        with self.assertRaisesRegexp(ValueError, "lvm snapshot origin must be a logical volume"):
-            LVMSnapShotDevice("snap1", parents=[vg], origin=pv)
-
-        with self.assertRaisesRegexp(ValueError, "only existing vorigin snapshots are supported"):
-            LVMSnapShotDevice("snap1", parents=[vg], vorigin=True)
-
-        lv.exists = True
-        snap1 = LVMSnapShotDevice("snap1", parents=[vg], origin=lv)
-
-        self.assertEqual(snap1.format, lv.format)
-        snap1.format = blivet.formats.getFormat("DM_snapshot_cow", exists=True)
-        self.assertEqual(snap1.format, lv.format)
-
-        self.assertEqual(snap1.isleaf, True)
-        self.assertEqual(snap1.direct, True)
-        self.assertEqual(lv.isleaf, False)
-        self.assertEqual(lv.direct, True)
-
-        self.assertEqual(snap1.dependsOn(lv), True)
-        self.assertEqual(lv.dependsOn(snap1), False)
-
-    def testLVMThinSnapShotDeviceInit(self):
-        pv = StorageDevice("pv1", fmt=blivet.formats.getFormat("lvmpv"),
-                           size=Size("1 GiB"))
-        vg = LVMVolumeGroupDevice("testvg", parents=[pv])
-        pool = LVMThinPoolDevice("pool1", parents=[vg], size=Size("500 MiB"))
-        thinlv = LVMThinLogicalVolumeDevice("thinlv", parents=[pool],
-                                            size=Size("200 MiB"))
-
-        with self.assertRaisesRegexp(ValueError, "lvm thin snapshots require an origin"):
-            LVMThinSnapShotDevice("snap1", parents=[pool])
-
-        with self.assertRaisesRegexp(ValueError, "lvm snapshot origin volume must already exist"):
-            LVMThinSnapShotDevice("snap1", parents=[pool], origin=thinlv)
-
-        with self.assertRaisesRegexp(ValueError, "lvm snapshot origin must be a logical volume"):
-            LVMThinSnapShotDevice("snap1", parents=[pool], origin=pv)
-
-        # now make the constructor succeed so we can test some properties
-        thinlv.exists = True
-        snap1 = LVMThinSnapShotDevice("snap1", parents=[pool], origin=thinlv)
-        self.assertEqual(snap1.isleaf, True)
-        self.assertEqual(snap1.direct, True)
-        self.assertEqual(thinlv.isleaf, True)
-        self.assertEqual(thinlv.direct, True)
-
-        self.assertEqual(snap1.dependsOn(thinlv), True)
-        self.assertEqual(thinlv.dependsOn(snap1), False)
-
-        # existing thin snapshots do not depend on their origin
-        snap1.exists = True
-        self.assertEqual(snap1.dependsOn(thinlv), False)
-
-    def testTargetSize(self):
-        pv = StorageDevice("pv1", fmt=blivet.formats.getFormat("lvmpv"),
-                           size=Size("1 GiB"))
-        vg = LVMVolumeGroupDevice("testvg", parents=[pv])
-        orig_size = Size("800 MiB")
-        lv = LVMLogicalVolumeDevice("testlv", parents=[vg], size=orig_size,
-                                    fmt=blivet.formats.getFormat("ext4"),
-                                    exists=True)
-
-        min_size = Size("200 MiB")
-        lv.format.exists = True
-        lv.format._minInstanceSize = min_size
-        lv.format._resizable = True
-
-        # Make sure things are as expected to begin with.
-        self.assertEqual(lv.minSize, min_size)
-        self.assertEqual(lv.maxSize, Size("1020 MiB"))
-        self.assertEqual(lv.size, orig_size)
-
-        # ValueError if size smaller than minSize
-        with self.assertRaisesRegexp(ValueError,
-                                     "size.*smaller than the minimum"):
-            lv.targetSize = Size("1 MiB")
-
-        # target size should be unchanged
-        self.assertEqual(lv.targetSize, orig_size)
-
-        # ValueError if size larger than maxSize
-        with self.assertRaisesRegexp(ValueError,
-                                     "size.*larger than the maximum"):
-            lv.targetSize = Size("1 GiB")
-
-        # target size should be unchanged
-        self.assertEqual(lv.targetSize, orig_size)
-
-        # successful set of target size should also be reflected in size attr
-        new_target = Size("900 MiB")
-        lv.targetSize = new_target
-        self.assertEqual(lv.targetSize, new_target)
-        self.assertEqual(lv.size, new_target)
-
-        # reset target size to original size
-        lv.targetSize = orig_size
-        self.assertEqual(lv.targetSize, orig_size)
-        self.assertEqual(lv.size, orig_size)
-
-class PartitionDeviceTestCase(unittest.TestCase):
-
-    def testTargetSize(self):
-        with sparsetmpfile("targetsizetest", Size("10 MiB")) as disk_file:
-            disk = DiskFile(disk_file)
-            disk.format = getFormat("disklabel", device=disk.path)
-            grain_size = Size(disk.format.alignment.grainSize)
-            sector_size = Size(disk.format.partedDevice.sectorSize)
-            start = int(grain_size)
-            orig_size = Size("6 MiB")
-            end = start + int(orig_size / sector_size) - 1
-            disk.format.addPartition(start, end)
-            partition = disk.format.partedDisk.getPartitionBySector(start)
-            self.assertNotEqual(partition, None)
-            self.assertEqual(orig_size, Size(partition.getLength(unit='B')))
-
-            device = PartitionDevice(os.path.basename(partition.path),
-                                     size=orig_size)
-            device.disk = disk
-            device.exists = True
-            device.partedPartition = partition
-
-            device.format = getFormat("ext4", device=device.path)
-            device.format.exists = True
-            # grain size should be 1 MiB
-            device.format._minInstanceSize = Size("2 MiB") + (grain_size / 2)
-            device.format._resizable = True
-
-            # Make sure things are as expected to begin with.
-            self.assertEqual(device.size, orig_size)
-            self.assertEqual(device.minSize, Size("3 MiB"))
-            # start sector's at 1 MiB
-            self.assertEqual(device.maxSize, Size("9 MiB"))
-
-            # ValueError if not Size
-            with self.assertRaisesRegexp(ValueError,
-                                         "new size must.*type Size"):
-                device.targetSize = 22
-
-            self.assertEqual(device.targetSize, orig_size)
-
-            # ValueError if size smaller than minSize
-            with self.assertRaisesRegexp(ValueError,
-                                         "size.*smaller than the minimum"):
-                device.targetSize = Size("1 MiB")
-
-            self.assertEqual(device.targetSize, orig_size)
-
-            # ValueError if size larger than maxSize
-            with self.assertRaisesRegexp(ValueError,
-                                         "size.*larger than the maximum"):
-                device.targetSize = Size("11 MiB")
-
-            self.assertEqual(device.targetSize, orig_size)
-
-            # ValueError if unaligned
-            with self.assertRaisesRegexp(ValueError, "new size.*not.*aligned"):
-                device.targetSize = Size("3.1 MiB")
-
-            self.assertEqual(device.targetSize, orig_size)
-
-            # successfully set a new target size
-            new_target = device.maxSize
-            device.targetSize = new_target
-            self.assertEqual(device.targetSize, new_target)
-            self.assertEqual(device.size, new_target)
-            parted_size = Size(device.partedPartition.getLength(unit='B'))
-            self.assertEqual(parted_size, device.targetSize)
-
-            # reset target size to original size
-            device.targetSize = orig_size
-            self.assertEqual(device.targetSize, orig_size)
-            self.assertEqual(device.size, orig_size)
-            parted_size = Size(device.partedPartition.getLength(unit='B'))
-            self.assertEqual(parted_size, device.targetSize)
-
-    def testMinMaxSizeAlignment(self):
-        with sparsetmpfile("minsizetest", Size("10 MiB")) as disk_file:
-            disk = DiskFile(disk_file)
-            disk.format = getFormat("disklabel", device=disk.path)
-            grain_size = Size(disk.format.alignment.grainSize)
-            sector_size = Size(disk.format.partedDevice.sectorSize)
-            start = int(grain_size)
-            end = start + int(Size("6 MiB") / sector_size)
-            disk.format.addPartition(start, end)
-            partition = disk.format.partedDisk.getPartitionBySector(start)
-            self.assertNotEqual(partition, None)
-
-            device = PartitionDevice(os.path.basename(partition.path))
-            device.disk = disk
-            device.exists = True
-            device.partedPartition = partition
-
-            # Typical sector size is 512 B.
-            # Default optimum alignment grain size is 2048 sectors, or 1 MiB.
-            device.format = getFormat("ext4", device=device.path)
-            device.format.exists = True
-            device.format._minInstanceSize = Size("2 MiB") + (grain_size / 2)
-            device.format._resizable = True
-
-            ##
-            ## minSize
-            ##
-
-            # The end sector based only on format min size should be unaligned.
-            min_sectors = int(device.format.minSize / sector_size)
-            min_end_sector = partition.geometry.start + min_sectors - 1
-            self.assertEqual(
-                disk.format.endAlignment.isAligned(partition.geometry,
-                                                   min_end_sector),
-                False)
-
-            # The end sector based on device min size should be aligned.
-            min_sectors = int(device.minSize / sector_size)
-            min_end_sector = partition.geometry.start + min_sectors - 1
-            self.assertEqual(
-                disk.format.endAlignment.isAligned(partition.geometry,
-                                                   min_end_sector),
-                True)
-
-            ##
-            ## maxSize
-            ##
-
-            # Add a partition starting three sectors past an aligned sector and
-            # extending to the end of the disk so that there's a free region
-            # immediately following the first partition with an unaligned end
-            # sector.
-            free = disk.format.partedDisk.getFreeSpaceRegions()[-1]
-            raw_start = int(Size("9 MiB") / sector_size)
-            start = disk.format.alignment.alignUp(free, raw_start) + 3
-            disk.format.addPartition(start, disk.partedDevice.length - 1)
-
-            # Verify the end of the free region immediately following the first
-            # partition is unaligned.
-            free = disk.format.partedDisk.getFreeSpaceRegions()[1]
-            self.assertEqual(disk.format.endAlignment.isAligned(free, free.end),
-                             False)
-
-            # The end sector based on device min size should be aligned.
-            max_sectors = int(device.maxSize / sector_size)
-            max_end_sector = partition.geometry.start + max_sectors - 1
-            self.assertEqual(
-                disk.format.endAlignment.isAligned(free, max_end_sector),
-                True)
-
-class DeviceNameTestCase(unittest.TestCase):
-    """Test device name validation"""
-
-    def testStorageDevice(self):
-        # Check that / and NUL are rejected along with . and ..
-        good_names = ['sda1', '1sda', 'good-name', 'cciss/c0d0']
-        bad_names = ['sda/1', 'sda\x00', '.', '..', 'cciss/..']
-
-        for name in good_names:
-            self.assertTrue(StorageDevice.isNameValid(name))
-
-        for name in bad_names:
-            self.assertFalse(StorageDevice.isNameValid(name))
-
-    def testVolumeGroup(self):
-        good_names = ['vg00', 'group-name', 'groupname-']
-        bad_names = ['-leading-hyphen', 'únicode', 'sp aces']
-
-        for name in good_names:
-            self.assertTrue(LVMVolumeGroupDevice.isNameValid(name))
-
-        for name in bad_names:
-            self.assertFalse(LVMVolumeGroupDevice.isNameValid(name))
-
-    def testLogicalVolume(self):
-        good_names = ['lv00', 'volume-name', 'volumename-']
-        bad_names = ['-leading-hyphen', 'únicode', 'sp aces',
-                     'snapshot47', 'pvmove0', 'sub_tmetastring']
-
-        for name in good_names:
-            self.assertTrue(LVMLogicalVolumeDevice.isNameValid(name))
-
-        for name in bad_names:
-            self.assertFalse(LVMLogicalVolumeDevice.isNameValid(name))
-
-class FakeNetDev(StorageDevice, NetworkStorageDevice):
-    _type = "fakenetdev"
-
-class NetDevMountOptionTestCase(unittest.TestCase):
-    def testNetDevSetting(self):
-        """ Verify netdev mount option setting after format assignment. """
-        netdev = FakeNetDev("net1")
-        dev = StorageDevice("dev1", fmt=getFormat("ext4"))
-        self.assertFalse("_netdev" in dev.format.options.split(","))
-
-        dev.parents.append(netdev)
-        dev.format = getFormat("ext4")
-        self.assertTrue("_netdev" in dev.format.options.split(","))
-
-    def testNetDevUpdate(self):
-        """ Verify netdev mount option setting after device creation. """
-        netdev = FakeNetDev("net1")
-        dev = StorageDevice("dev1", fmt=getFormat("ext4"))
-        self.assertFalse("_netdev" in dev.format.options.split(","))
-
-        dev.parents.append(netdev)
-
-        # these create methods shouldn't write anything to disk
-        netdev.create()
-        dev.create()
-
-        self.assertTrue("_netdev" in dev.format.options.split(","))
-
-
 if __name__ == "__main__":
     unittest.main()
-
